worker/requirements.txt
sqlalchemy
asyncpg
greenlet
apscheduler
trafilatura
beautifulsoup4
httpx
lxml
feedparser
python-dateutil
pydantic
pydantic-settings
playwright
